requirements.txt
fastapi
uvicorn[standard]
jinja2
google-cloud-firestore
google-cloud-storage
httpx
beautifulsoup4
PyYAML
pydantic
